+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Comment block (`{{# }}`) support. (#47)
 - Delegate parsing of the main body of a `.vto` file to the JetBrains HTML plugin. (#51)
 - Variable block with (`{{ }}`) and wihtout trimming (`{{- -}}`) support. (#36) (#38)
+- Variable pipe (`{{ expression |> call }}`) support (#39)
 
 ### Fixed